+++ core/api_container/execution/exit_codes/exit_codes.go
@@ -1,0 +1,9 @@
+package exit_codes
+
+const (
+	TestCompletedInTimeoutExitCode = 0
+	NoTestSuiteRegisteredExitCode = 1
+	TestHitTimeoutExitCode = 2
+	ShutdownSignalExitCode = 3
+	OutOfOrderTestStatusExitCode = 4
+)