+++ core/todo_rename_new_initializer/main.go
@@ -3,24 +3,50 @@ package main
 import (
 	"bufio"
 	"context"
+	"encoding/binary"
 	"flag"
 	"fmt"
 	"github.com/docker/docker/client"
 	"github.com/docker/go-connections/nat"
+	"github.com/google/uuid"
 	"github.com/kurtosis-tech/kurtosis/api_container/api"
+	"github.com/kurtosis-tech/kurtosis/api_container/execution/exit_codes"
 	"github.com/kurtosis-tech/kurtosis/commons/docker"
 	"github.com/kurtosis-tech/kurtosis/commons/networks"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
 	"io/ioutil"
+	"math"
+	"net"
 	"os"
+	"strings"
 	"time"
 )
 
 const (
+	successExitCode = 0
+	failureExitCode = 1
+
 	testNamesFilepath = "/shared/test-names-filepath"
-	kurtosisApiStopContainerTimeout = 30 * time.Second
 	dockerSocket = "/var/run/docker.sock"
+	testNameArgSeparator = ","
+
+	// This is the IP address that the first Docker subnet will be doled out from, with subsequent Docker networks doled out with
+	//  increasing IPs corresponding to the NETWORK_WIDTH_BITS
+	subnetStartAddr = "172.23.0.0"
+	betsInIPv4Addr  = 32
+
+	bridgeNetworkName = "bridge"
+
+	// When we're tearing down the network after running a test, how long we'll give each contaienr to gracefully stop itself
+	//  before hard-killing it
+	networkTeardownContainerStopTimeout = 30 * time.Second
+
+	// TODO Parameterize this! (will require getting information from testsuite)
+	defaultNetworkWidthBits = 8
+
+	// TODO Parameterize this!!
+	testVolumeMountLocation = "/shared"
 
 	// Test suite environment variables
 	testNamesFilepathEnvVar    = "TEST_NAMES_FILEPATH"
@@ -35,11 +61,12 @@ const (
 	gatewayIpEnvVar            = "GATEWAY_IP"
 	logLevelEnvVar             = "LOG_LEVEL"
 	apiLogFilepathEnvVar       = "LOG_FILEPATH"
+	apiContainerIpAddrEnvVar   = "API_CONTAINER_IP"
+	testSuiteContainerIpAddrEnvVar   = "TEST_SUITE_CONTAINER_IP"
 
 
 	// TODO parameterize
 	kurtosisApiImage        = "kurtosistech/kurtosis-core_api"
-	bridgeNetworkId         = "b453ce4bac01"
 	bridgeNetworkSubnetMask = "172.17.0.0/16"
 	bridgeNetworkGatewayIp  = "172.17.0.1"
 )
@@ -49,61 +76,87 @@ func main() {
 		"test-suite-image",
 		"",
 		"The name of the Docker image of the test suite that will be run")
+	testNamesArg := flag.String(
+		"test-names",
+		"",
+		"List of test names to run, separated by '" + testNameArgSeparator + "' (default or empty: run all tests)",
+	)
+	// TODO add a "list tests" flag
 	flag.Parse()
+
+	// TODO make this configurable
+	logrus.SetLevel(logrus.TraceLevel)
 
 	dockerClient, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
 	if err != nil {
 		logrus.Errorf("An error occurred creating the Docker client: %v", err)
-		os.Exit(1)
+		os.Exit(failureExitCode)
 	}
 
 	dockerManager, err := docker.NewDockerManager(logrus.StandardLogger(), dockerClient)
 	if err != nil {
 		logrus.Errorf("An error occurred creating the Docker manager: %v", err)
-		os.Exit(1)
+		os.Exit(failureExitCode)
 	}
 
-	freeIpAddrTracker, err := networks.NewFreeIpAddrTracker(
-		logrus.StandardLogger(),
-		bridgeNetworkSubnetMask,
-		map[string]bool{
-			bridgeNetworkGatewayIp: true,	// gateway IP
-		})
+	testNamesToRun, err := getTestNamesToRun(*testNamesArg, *testSuiteImageArg, dockerManager)
 	if err != nil {
-		logrus.Errorf("An error occurred creating the free IP address tracker: %v", err)
-		os.Exit(1)
-	}
-
-	// TODO actually use tests
-	_, err = getTests(*testSuiteImageArg, dockerManager, freeIpAddrTracker)
-	if err != nil {
-		logrus.Errorf("An error occurred getting the names of the tests in the test suite: %v", err)
-		os.Exit(1)
-	}
-
-	// TODO parameterize test name
-	testPassed, err := runTest(*testSuiteImageArg, dockerManager, freeIpAddrTracker, "FOO")
-	if err != nil {
-		logrus.Errorf("An error occurred that prevented retrieving test pass/fail status:")
+		logrus.Errorf("An error occurred when validating the list of tests to run:")
 		fmt.Fprintln(logrus.StandardLogger().Out, err)
-		os.Exit(1)
+		os.Exit(failureExitCode)
 	}
 
-	if !testPassed {
-		// TODO insert test name
-		logrus.Error("The test failed")
-		os.Exit(1)
-	} else {
-		// TODO insert test name
-		logrus.Info("The test passed")
-		os.Exit(0)
+	subnetStartIpInt, err := getSubnetStartIpInt()
+	if err != nil {
+		logrus.Errorf("An error getting the subnet start IP integer:")
+		fmt.Fprintln(logrus.StandardLogger().Out, err)
+		os.Exit(failureExitCode)
 	}
+
+	executionId := uuid.New()
+
+	// TODO sort these test names so we're doing things deterministically
+	// TODO Parallelize these bitches
+	allTestsPassed := true
+	testIndex := 0
+	for testName, _ := range testNamesToRun {
+		testPassed, testExecutionErr := runSingleTest(
+			executionId,
+			*testSuiteImageArg,
+			dockerManager,
+			testName,
+			subnetStartIpInt,
+			testIndex)
+		if testExecutionErr != nil {
+			logrus.Errorf("An error occurred that prevented retrieving test pass/fail status for test '%v':", testName)
+			fmt.Fprintln(logrus.StandardLogger().Out, testExecutionErr)
+			logrus.Errorf("%v: ERRORED", testName)
+		} else {
+			if testPassed {
+				logrus.Infof("%v: PASSED", testName)
+			} else {
+				logrus.Errorf("%v: FAILED", testName)
+			}
+		}
+		allTestsPassed = allTestsPassed && (testExecutionErr == nil && testPassed)
+		testIndex++
+	}
+
+	var exitCode int
+	if allTestsPassed {
+		exitCode = successExitCode
+	} else {
+		exitCode = failureExitCode
+	}
+	os.Exit(exitCode)
 }
 
 /*
 Spins up a testsuite container in test-listing mode and reads the tests that it spits out
  */
-func getTests(testSuiteImage string, dockerManager *docker.DockerManager, freeIpAddrTracker *networks.FreeIpAddrTracker) (map[string]bool, error) {
+func getAllTestNames(
+			testSuiteImage string,
+			dockerManager *docker.DockerManager) (map[string]bool, error) {
 	// Create the tempfile that the testsuite image will write test names to
 	tempFp, err := ioutil.TempFile("", "test-names")
 	if err != nil {
@@ -111,17 +164,24 @@ func getTests(testSuiteImage string, dockerManager *docker.DockerManager, freeIp
 	}
 	tempFp.Close()
 
-	testListingContainerIp, err := freeIpAddrTracker.GetFreeIpAddr()
+	bridgeNetworkIds, err := dockerManager.GetNetworkIdsByName(context.Background(), bridgeNetworkName)
 	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred getting a free IP address for the testsuite container")
+		return nil, stacktrace.Propagate(err, "An error occurred gettings the network IDs matching the '%v' network", bridgeNetworkName)
 	}
+	if len(bridgeNetworkIds) == 0 || len(bridgeNetworkIds) > 1 {
+		return nil, stacktrace.NewError(
+			"%v Docker network IDs were returned for the '%v' network - this is very strange!",
+			len(bridgeNetworkIds),
+			bridgeNetworkName)
+	}
+	bridgeNetworkId := bridgeNetworkIds[0]
 
 	testListingContainerId, err := dockerManager.CreateAndStartContainer(
 		context.Background(),
 		testSuiteImage,
 		// TODO parameterize these
 		bridgeNetworkId,
-		testListingContainerIp,
+		nil,  // Nil because the bridge network will assign IPs on its own; we don't need to (and won't know what IPs are already used)
 		map[nat.Port]bool{},
 		nil,
 		map[string]string{
@@ -163,12 +223,90 @@ func getTests(testSuiteImage string, dockerManager *docker.DockerManager, freeIp
 	return testNames, nil
 }
 
-// TODO make output status  more meaningful
-/*
-Runs the given test by spinning up test suite & Kurtosis API containers
- */
-func runTest(testSuiteImage string, dockerManager *docker.DockerManager, freeIpAddrTracker *networks.FreeIpAddrTracker, test string) (bool, error){
-	// TODO segregate these into their own subnet
+func getTestNamesToRun(
+			testsToRunStr string,
+			testSuiteImage string,
+			dockerManager *docker.DockerManager) (map[string]bool, error) {
+	// Split user-input string into actual candidate test names
+	testNamesArgStr := strings.TrimSpace(testsToRunStr)
+	testNamesToRun := map[string]bool{}
+	if len(testNamesArgStr) > 0 {
+		testNamesList := strings.Split(testNamesArgStr, testNameArgSeparator)
+		for _, name := range testNamesList {
+			testNamesToRun[name] = true
+		}
+	}
+
+	allTestNames, err := getAllTestNames(testSuiteImage, dockerManager)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred getting the names of the tests in the test suite")
+	}
+
+	// If the user doesn't specify any test names to run, do all of them
+	if len(testNamesToRun) == 0 {
+		testNamesToRun = map[string]bool{}
+		for testName := range allTestNames {
+			testNamesToRun[testName] = true
+		}
+	}
+
+	// Validate all the requested tests exist
+	for testName := range testNamesToRun {
+		if _, found := allTestNames[testName]; !found {
+			return nil, stacktrace.NewError("No test registered with name '%v'", testName)
+		}
+	}
+	return testNamesToRun, nil
+}
+
+
+func runSingleTest(
+		executionId uuid.UUID,
+		testSuiteImage string,
+		dockerManager *docker.DockerManager,
+		testName string,
+		subnetStartIpInt uint32,
+		testIndex int) (bool, error){
+
+	// TODO pull netwokrwidthbits from the testsuite
+	subnetMask, err := getTestSubnetMask(subnetStartIpInt, defaultNetworkWidthBits, testIndex)
+	if err != nil {
+		return false, stacktrace.Propagate(err, "An error occurred getting subnet mask for test '%v':", testName)
+	}
+
+	uniqueTestIdentifier := fmt.Sprintf("%v-%v", executionId.String(), testName)
+
+	volumeName := uniqueTestIdentifier
+	logrus.Debugf("Creating Docker volume %v which will be shared with the test network...", volumeName)
+	if err := dockerManager.CreateVolume(context.Background(), volumeName); err != nil {
+		return false, stacktrace.Propagate(err, "Error creating Docker volume to share amongst test nodes for test %v", testName)
+	}
+	logrus.Debugf("Docker volume %v created successfully", volumeName)
+
+	logrus.Infof("Creating Docker network for test with subnet mask %v...", subnetMask)
+	freeIpAddrTracker, err := networks.NewFreeIpAddrTracker(
+		logrus.StandardLogger(),
+		subnetMask,
+		map[string]bool{})
+	if err != nil {
+		return false, stacktrace.Propagate(err, "An error occurred creating the free IP address tracker for test %v", testName)
+	}
+	gatewayIp, err := freeIpAddrTracker.GetFreeIpAddr()
+	if err != nil {
+		return false, stacktrace.Propagate(err, "An error occurred getting a free IP for the gateway for test %v", testName)
+	}
+	networkName := fmt.Sprintf("%v-%v", executionId.String(), testName)
+	// TODO different context for parallelization?
+	networkId, err := dockerManager.CreateNetwork(context.Background(), networkName, subnetMask, gatewayIp)
+	if err != nil {
+		return false, stacktrace.Propagate(err, "Error occurred creating Docker network %v for test %v", networkName, testName)
+	}
+	defer removeNetworkDeferredFunc(dockerManager, networkId)
+	logrus.Infof("Docker network %v created successfully", networkId)
+
+	// TODO use hostnames rather than IPs, which makes things nicer and which we'll need for Docker swarm support
+	// We need to create the IP addresses for BOTH containers because the testsuite needs to know the IP of the API
+	//  container which will only be started after the testsuite container
 	kurtosisApiIp, err := freeIpAddrTracker.GetFreeIpAddr()
 	if err != nil {
 		return false, stacktrace.Propagate(err, "An error occurred getting an IP for the Kurtosis API container")
@@ -178,34 +316,42 @@ func runTest(testSuiteImage string, dockerManager *docker.DockerManager, freeIpA
 		return false, stacktrace.Propagate(err, "An error occurred getting an IP for the test suite container running the test")
 	}
 
+	logrus.Debugf(
+		"Test suite container IP: %v; kurtosis API container IP: %v",
+		testRunningContainerIp.String(),
+		kurtosisApiIp.String())
+
+	logrus.Infof("Creating test suite container that will run the test...")
 	testRunningContainerId, err := dockerManager.CreateAndStartContainer(
 		context.Background(),
 		testSuiteImage,
 		// TODO parameterize network stuff
-		bridgeNetworkId,
+		networkId,
 		testRunningContainerIp,
 		map[nat.Port]bool{},
 		nil,
 		map[string]string{
 			testNamesFilepathEnvVar: "", // We leave this blank because we want test execution, not listing
-			testEnvVar:          test,                  // We leave this blank to signify that we want test listing, not test execution
+			testEnvVar:          testName,                  // We leave this blank to signify that we want test listing, not test execution
 			kurtosisApiIpEnvVar: kurtosisApiIp.String(), // Because we're doing test listing, this can be blank
-			// TODO pipe this to a proper volume location
-			testSuiteLogFilepathEnvVar: "/tmp/log.txt",
+			testSuiteLogFilepathEnvVar: testVolumeMountLocation + "/test.log",
 		},
 		map[string]string{},
-		map[string]string{})
+		map[string]string{
+			volumeName: testVolumeMountLocation,
+		})
 	if err != nil {
 		return false, stacktrace.Propagate(err, "An error occurred creating the test suite container to run the test")
 	}
+	logrus.Info("Successfully created test suite container to run the test")
 
-	// TODO replace this with actually running tests
+	logrus.Info("Creating Kurtosis API container...")
 	kurtosisApiPort := nat.Port(fmt.Sprintf("%v/tcp", api.KurtosisAPIContainerPort))
 	kurtosisApiContainerId, err := dockerManager.CreateAndStartContainer(
 		context.Background(),
 		// TODO parameterize these
 		kurtosisApiImage,
-		bridgeNetworkId,
+		networkId,
 		kurtosisApiIp,
 		map[nat.Port]bool{
 			kurtosisApiPort: true,
@@ -213,48 +359,118 @@ func runTest(testSuiteImage string, dockerManager *docker.DockerManager, freeIpA
 		nil,
 		map[string]string{
 			testSuiteContainerIdEnvVar: testRunningContainerId,
-			networkIdEnvVar:            bridgeNetworkId,
-			subnetMaskEnvVar: bridgeNetworkSubnetMask,
-			gatewayIpEnvVar: bridgeNetworkGatewayIp,
-			// TODO change this
-			logLevelEnvVar:             "trace",
-			apiLogFilepathEnvVar: "/tmp/logfile",
+			// TODO Change all of these
+			networkIdEnvVar: networkId,
+			subnetMaskEnvVar: subnetMask,
+			gatewayIpEnvVar: gatewayIp.String(),
+			// TODO make this parameterizable
+			logLevelEnvVar: "trace",
+			apiLogFilepathEnvVar: testVolumeMountLocation + "/api.log",
+			apiContainerIpAddrEnvVar: kurtosisApiIp.String(),
+			testSuiteContainerIpAddrEnvVar: testRunningContainerIp.String(),
 		},
 		map[string]string{
 			dockerSocket: dockerSocket,
 		},
-		map[string]string{})
+		map[string]string{
+			volumeName: testVolumeMountLocation,
+		})
 	if err != nil {
 		return false, stacktrace.Propagate(err, "An error occurred creating the Kurtosis API container")
 	}
+	logrus.Infof("Successfully created Kurtosis API container")
 
-	// The Kurtosis API will be our indication of whether the test suite container stopped within the timeout or not
 	// TODO add a timeout waiting for Kurtosis API container to stop???
+	// The Kurtosis API will be our indication of whether the test suite container stopped within the timeout or not
+	logrus.Info("Waiting for Kurtosis API container to exit...")
 	kurtosisApiExitCode, err := dockerManager.WaitForExit(
 		context.Background(),
 		kurtosisApiContainerId)
 	if err != nil {
-		logrus.Errorf("An error occurred waiting for the exit of the Kurtosis API container: %v", err)
+		return false, stacktrace.Propagate(err, "An error occurred waiting for the exit of the Kurtosis API container: %v", err)
 	}
 
-	if kurtosisApiExitCode != 0 {
-		// TODO change this with tearing down the entire network
-		// The testsuite container didn't stop within the tiemout; make a best-effort attempt to stop the testsuite container
-		if err := dockerManager.StopContainer(context.Background(), testRunningContainerId, 30 * time.Second); err != nil {
-			logrus.Error("An error occurred during our best-effort attempt at stopping the testsuite container which exceeded its test timeout:")
-			fmt.Fprintln(logrus.StandardLogger().Out, err)
-		}
-		return false, stacktrace.NewError("The testsuite container didn't stop within the hard test timeout")
+	var testStatusRetrievalError error
+	switch kurtosisApiExitCode {
+	case exit_codes.TestCompletedInTimeoutExitCode:
+		testStatusRetrievalError = nil
+	case exit_codes.OutOfOrderTestStatusExitCode:
+		testStatusRetrievalError = stacktrace.NewError("The Kurtosis API container received an out-of-order " +
+			"test execution status update; this is a Kurtosis code bug")
+	case exit_codes.TestHitTimeoutExitCode:
+		testStatusRetrievalError = stacktrace.NewError("The test failed to complete within the hard test " +
+			"timeout (test_execution_timeout + setup_buffer)")
+	case exit_codes.NoTestSuiteRegisteredExitCode:
+		testStatusRetrievalError = stacktrace.NewError("The test suite failed to register itself with the " +
+			"Kurtosis API container; this is a bug with the test suite")
+	case exit_codes.ShutdownSignalExitCode:
+		testStatusRetrievalError = stacktrace.NewError("The Kurtosis API container exited due to receiving " +
+			"a shutdown signal; if this is not expected, it's a Kurtosis bug")
+	default:
+		testStatusRetrievalError = stacktrace.NewError("The Kurtosis API container exited with an unrecognized " +
+			"exit code %v; this is a code bug in Kurtosis indicating that the new exit code needs to be mapped " +
+			"for the initializer", kurtosisApiExitCode)
 	}
+	if testStatusRetrievalError != nil {
+		logrus.Error("An error occurred that prevented retrieval of the test completion status")
+		return false, testStatusRetrievalError
+	}
+	logrus.Info("The test suite container running the test exited within the hard test timeout")
 
-	// Test stopped within the timeout; examine the testsuite container for actual test result
-	testRunningExitCode, err := dockerManager.WaitForExit(
+	// The test suite container will already have stopped, so now we get the exit code
+	testSuiteExitCode, err := dockerManager.WaitForExit(
 		context.Background(),
 		testRunningContainerId)
 	if err != nil {
-		return false, stacktrace.Propagate(
-			err,
-			"The testsuite container running the test stopped within the timeout, but an error occurred retrieving the exit code")
+		return false, stacktrace.Propagate(err, "An error occurred retrieving the test suite container exit code")
 	}
-	return testRunningExitCode == 0, nil
+	return testSuiteExitCode == 0, nil
+}
+
+func getSubnetStartIpInt() (uint32, error) {
+	subnetStartIp := net.ParseIP(subnetStartAddr)
+	if subnetStartIp == nil {
+		return 0, stacktrace.NewError("Subnet start IP %v was not a valid IP address; this is a code problem", subnetStartAddr)
+	}
+
+	// The IP can be either 4 bytes or 16 bytes long; we need to handle both
+	//  else we'll get a silent 0 value for the int!
+	// See https://gist.github.com/ammario/649d4c0da650162efd404af23e25b86b
+	var subnetStartIpInt uint32
+	if len(subnetStartIp) == 16 {
+		subnetStartIpInt = binary.BigEndian.Uint32(subnetStartIp[12:16])
+	} else {
+		subnetStartIpInt = binary.BigEndian.Uint32(subnetStartIp)
+	}
+	return subnetStartIpInt, nil
+}
+
+func getTestSubnetMask(subnetStartIpInt uint32, networkWidthBits uint32, testIndex int) (string, error) {
+	subnetMaskBits := betsInIPv4Addr - networkWidthBits
+
+	// Pick the next free available subnet IP, considering all the tests we've started previously
+	subnetIpInt := subnetStartIpInt + uint32(testIndex) * uint32(math.Pow(2, float64(networkWidthBits)))
+	subnetIp := make(net.IP, 4)
+	binary.BigEndian.PutUint32(subnetIp, subnetIpInt)
+	subnetCidrStr := fmt.Sprintf("%v/%v", subnetIp.String(), subnetMaskBits)
+
+	return subnetCidrStr, nil
+
+}
+
+/*
+Helper function for making a best-effort attempt at removing a network and logging any error states; intended to be run
+as a deferred function.
+*/
+func removeNetworkDeferredFunc(dockerManager *docker.DockerManager, networkId string) {
+	logrus.Infof("Attempting to remove Docker network with id %v...", networkId)
+	// We use the background context here because we want to try and tear down the network even if the context the test was running in
+	//  was cancelled. This might not be right - the right way to do it might be to pipe a separate context for the network teardown to here!
+	if err := dockerManager.RemoveNetwork(context.Background(), networkId, networkTeardownContainerStopTimeout); err != nil {
+		logrus.Errorf("An error occurred removing Docker network with ID %v:", networkId)
+		logrus.Error(err.Error())
+		logrus.Error("NOTE: This means you will need to clean up the Docker network manually!!")
+	} else {
+		logrus.Infof("Docker network with ID %v successfully removed", networkId)
+	}
 }