+++ core/CHANGELOG.md
@@ -4,6 +4,7 @@
 * Build a v0 of Docker container containing the Kurtosis API 
 * Add registration endpoint to the API container
 * Fix bugs with registration endpoint in API container
+* Upgrade new initializer to actually run a test suite!
 
 # 0.9.0
 * Change ConfigurationID to be a string